feat: Add day 21 part 2
This is a very bad performing solution that theoretically should work.

+++ module-tool/src/main/kotlin/de/twomartens/adventofcode/day21/graph/Graph.kt
@@ -10,7 +10,7 @@ class Graph(val rows: List<List<Boolean>>, val startPosition: Index) {
                 val fields = line.split("").filterNot { it.isBlank() }
                 val tiles = fields.mapIndexed { columnIndex, it ->
                     if (it == "S") {
-                        startPosition = Index(rowIndex, columnIndex)
+                        startPosition = Index(rowIndex.toLong(), columnIndex.toLong())
                     }
                     it == "." || it == "S"
                 }
